Move consistency_* functions into sweettest module
