+++ crates/holochain/src/sweettest/mod.rs
@@ -20,6 +20,7 @@ mod sweet_conductor_batch;
 mod sweet_conductor_config;
 mod sweet_conductor_config_rendezvous;
 mod sweet_conductor_handle;
+pub mod sweet_consistency;
 mod sweet_dna;
 mod sweet_network;
 mod sweet_zome;
@@ -32,6 +33,7 @@ pub use sweet_conductor_batch::*;
 pub use sweet_conductor_config::*;
 pub use sweet_conductor_config_rendezvous::*;
 pub use sweet_conductor_handle::*;
+pub use sweet_consistency::*;
 pub use sweet_dna::*;
 pub use sweet_network::*;
 pub use sweet_zome::*;
